Added : Refactored the code

+++ src/main/java/com/example/demo/service/EmployeePayrollService.java
@@ -1,7 +1,7 @@
 package com.example.demo.service;
-
-import com.example.demo.dto.EmployeePayrollDTO;
-import com.example.demo.model.EmployeePayrollData;
+import com.example.demo.dto.EmployeeDTO;
+import com.example.demo.exception.EmployeePayrollException;
+import com.example.demo.model.Employee;
 import com.example.demo.repository.EmployeePayrollRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,54 +10,54 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
+
+//Created EmployeePayrollService class to serve api calls done by controller layer
 public class EmployeePayrollService implements IEmployeePayrollService {
+
+    //Autowired EmployeePayrollRepository interface to inject its dependency here
     @Autowired
     EmployeePayrollRepository repository;
 
-    public String getMessage(String name) {
-        return "Welcome To Employee Program " + name;
-    }
+//    public String getWelcome() {
+//        return "Welcome to Employee Payroll !!!";
+//    }
 
-    public String postMessage(EmployeePayrollDTO employee) {
-        return "Hello Employee " + employee.getFirstName() + "" + employee.getLastName() + "!";
-    }
-
-    public String putMessage(String name) {
-        return "Hey Dude , " + name;
-    }
-
-    public String getWelcome() {
-        return "Welcome to Employee Payroll App.....!";
-    }
-
-    @Override
-    public EmployeePayrollData postDataToRepo(EmployeePayrollDTO employeepayrolldata) {
-        EmployeePayrollData newEmployee = new EmployeePayrollData(employeepayrolldata);
+    public Employee postDataToRepo(EmployeeDTO employeeDTO) {
+        Employee newEmployee = new Employee(employeeDTO);
         repository.save(newEmployee);
         return newEmployee;
     }
 
-    @Override
-    public List<EmployeePayrollData> getAllData() {
-        List<EmployeePayrollData> list = repository.findAll();
+    public List<Employee> getAllData() {
+        List<Employee> list = repository.findAll();
         return list;
     }
 
-    @Override
-    public Optional<EmployeePayrollData> getDataById(Integer id) {
-        Optional<EmployeePayrollData> newEmployee = repository.findById(id);
-        return newEmployee;
+    public Employee getDataById(Integer id) {
+        Optional<Employee> newEmployee = repository.findById(id);
+        if (newEmployee.isPresent()) {
+            return newEmployee.get();
+        } else throw new EmployeePayrollException("Employee id not found");
     }
 
-    public EmployeePayrollData updateDataById(Integer id, EmployeePayrollDTO employeepayrollDTO) {
-        EmployeePayrollData newEmployee = new EmployeePayrollData(id, employeepayrollDTO);
-        repository.save(newEmployee);
-        return newEmployee;
+    public Employee updateDataById(Integer id, EmployeeDTO employeeDTO) {
+        Optional<Employee> newEmployee = repository.findById(id);
+        if (newEmployee.isPresent()) {
+            Employee employee = new Employee(id, employeeDTO);
+            repository.save(employee);
+            return employee;
+        } else {
+            throw new EmployeePayrollException("Employee Not found");
+        }
     }
 
     public String deleteDataById(Integer id) {
-        repository.deleteById(id);
-        return "Employee with ID:" + id + " deleted";
+        Optional<Employee> newEmployee = repository.findById(id);
+        if (newEmployee.isEmpty()) {
+            throw new EmployeePayrollException("Employee Details not found");
+        } else {
+            repository.deleteById(id);
+        }
+        return null;
     }
-
 }
